+++ bliss/config/conductor/log_server.py
@@ -1,7 +1,6 @@
 import sys
 import pickle
 import logging
-import logging.handlers
 from logging.handlers import RotatingFileHandler
 import struct
 import argparse
@@ -10,53 +9,113 @@ import pathlib
 from gevent.server import StreamServer
 
 
-INITIALIZED_SESSIONS = []  # store names of initialized sessions for creating FileHandlers
-
 _log = logging.getLogger("log_server")
 
 
-class Handler:
-    def __init__(self, db_path, log_size):
+class LogServer(StreamServer):
+    """Process logging record received inside the `handle` function,
+    and dispatch them into different logging handlers."""
+
+    _handlers = {}
+    """
+    Store initialized FileHandlers.
+
+    There is a single handler per session and application.
+    """
+
+    def __init__(self, listener, db_path, log_size):
         self.db_path = db_path
         self.log_size = log_size
+        super(LogServer, self).__init__(listener, self.handle)
 
-    def __call__(self, socket, address):
+    def handle(self, socket, address):
         while True:
-            chunk = socket.recv(4)
-            if len(chunk) < 4:
-                break
-            slen = struct.unpack(">L", chunk)[0]
-            chunk = socket.recv(slen)
-            if not len(chunk):  # socket closed by client
-                break
+            try:
+                chunk = socket.recv(4)
+                if len(chunk) < 4:
+                    break
+                slen = struct.unpack(">L", chunk)[0]
+                chunk = socket.recv(slen)
+                if not len(chunk):  # socket closed by client
+                    break
 
-            while len(chunk) < slen:
-                data = socket.recv(slen - len(chunk))
-                if not len(data):  # socket closed by client
-                    return
-                chunk = chunk + data
-            obj = pickle.loads(chunk)
+                while len(chunk) < slen:
+                    data = socket.recv(slen - len(chunk))
+                    if not len(data):  # socket closed by client
+                        return
+                    chunk = chunk + data
+                record_dict = pickle.loads(chunk)
 
-            check_session(obj["session"], self.db_path, self.log_size)
+                # Backward compatibility with BLISS <1.6
+                # FIXME: This can be remove for BLISS 1.7
+                if "application" not in record_dict:
+                    record_dict["application"] = "bliss"
 
-            # adding a 'session' argument to log messages
-            record = logging.makeLogRecord(obj)
-            handle_log_record(record)
+                record = logging.makeLogRecord(record_dict)
+                self.log_record(record)
+            except Exception:
+                _log.error("Error while processing message", exc_info=True)
+                raise
 
+    def get_handler(self, record):
+        """Returns the dedicated handler associated to this record
 
-def handle_log_record(record):
-    logger = logging.getLogger(record.name)
-    logger.handle(record)
+        If the handle is not yet exists, it is created.
+        """
+        session = record.session
+        application = record.application
+        key = session, application
+        if key not in self._handlers:
+            handler = self.create_handler(record)
+            # This only create a single time handler anyway it is None
+            self._handlers[key] = handler
+        else:
+            handler = self._handlers[key]
+        return handler
 
+    def create_handler(self, record):
+        """
+        Create a new handler associated to this record.
 
-def check_session(session, root_path, log_size=1):
-    """
-    Check if a session has a RotatingFileHandler and if not it
-    creates one
-    """
-    if session not in INITIALIZED_SESSIONS:
-        # creating handler
-        dir_path = pathlib.Path(root_path)
+        If will create a dedicated RotatingFileHandler per tuple session/application.
+
+        If the record is not valid None is returned.
+        """
+        session = record.session
+        application = record.application
+        if application == "bliss":
+            handler = self.create_bliss_session_handler(
+                self.db_path, session, self.log_size
+            )
+        elif application == "flint":
+            handler = self.create_flint_session_handler(
+                self.db_path, session, self.log_size
+            )
+        else:
+            _log.error(
+                "Unknown application '%s'. Logs from this source will be ignored.",
+                application,
+            )
+            handler = None
+        return handler
+
+    def log_record(self, record):
+        """Log a record to the beacon log service
+
+        Arguments:
+            record: A logging.LogRecord object with extra attributes `session`
+                and `application`.
+        """
+        handler = self.get_handler(record)
+        if handler is not None:
+            handler.emit(record)
+
+    def create_bliss_session_handler(self, dir_path, session, log_size):
+        """Create a dedicated handler to store log in a file for a specific
+        BLISS session"""
+
+        # create root directory if needed
+        dir_path = pathlib.Path(dir_path)
         if not dir_path.exists():
             dir_path.mkdir()
 
@@ -66,21 +125,35 @@ def check_session(session, root_path, log_size=1):
             file_path, maxBytes=1024 ** 2 * log_size, backupCount=10
         )
 
-        def filter_func(rec):
-            # filter if a message is for the right session
-            # and consequently the right log file
-            if rec.session == session:
-                return True
-            return False
-
         formatter = logging.Formatter(
             "%(asctime)s %(session)s %(name)s %(levelname)s : %(msg)s"
         )  # adapt to needs
-        handler.addFilter(filter_func)
         handler.setFormatter(formatter)
-        logging.getLogger().addHandler(handler)
+        return handler
 
-        INITIALIZED_SESSIONS.append(session)
+    def create_flint_session_handler(self, dir_path, session, log_size):
+        """Create a dedicated handler to store log in a file for Flint when
+        running in a specific session.
+
+        As many Flint can run at same time. The process identifier is also
+        logged.
+        """
+        # Create root directory if needed
+        dir_path = pathlib.Path(dir_path)
+        if not dir_path.exists():
+            dir_path.mkdir()
+
+        file_path = dir_path / f"flint_{session}.log"
+        _log.info(f"Appending to file {file_path}")
+        handler = RotatingFileHandler(
+            file_path, maxBytes=1024 ** 2 * log_size, backupCount=10
+        )
+
+        formatter = logging.Formatter(
+            "%(asctime)s %(session)s %(process)s %(name)s %(levelname)s : %(msg)s"
+        )  # adapt to needs
+        handler.setFormatter(formatter)
+        return handler
 
 
 def main(args=None):
@@ -100,12 +173,12 @@ def main(args=None):
         "--log-output-folder", "--log_output_folder", type=str, dest="log_output_folder"
     )
     p.add_argument("--log-size", "--log_size", type=float, dest="log_size")
-    _options = p.parse_args(args)
+    options = p.parse_args(args)
 
-    _log.info(f"Initialize StreamServer on port {_options.port}")
-    server = StreamServer(
-        ("0.0.0.0", _options.port),
-        Handler(_options.log_output_folder, _options.log_size),
+    _log.info(f"Initialize LogServer on port {options.port}")
+
+    server = LogServer(
+        ("0.0.0.0", options.port), options.log_output_folder, options.log_size
     )
     server.serve_forever()